+++ spec/commands/draw_spec.rb
@@ -195,4 +195,17 @@ describe SlackGamebot::Commands::Draw do
       end
     end
   end
+
+  context 'with a proposed challenge' do
+    let(:challenged) { Fabricate(:user, user_name: 'username', channel: channel) }
+    let!(:challenge) { Fabricate(:challenge, challenged: [challenged], channel: channel) }
+
+    it 'cannot draw a challenge that has not been accepted' do
+      expect(message: '@gamebot draw', user: challenged.user_id, channel: challenge.channel).to respond_with_slack_message(
+        'No challenge to draw!'
+      )
+      expect(challenge.reload.state).to eq ChallengeState::PROPOSED
+      expect(Match.count).to eq 0
+    end
+  end
 end